Apartments Data

Datasets apartments and apartments_test are artificial, generated form the same model. Structure of the dataset is copied from real dataset from PBImisc package, but they were generated in a way to mimic effect of Anscombe quartet for complex black box models.

data(apartments)

Format

a data frame with 1000 rows and 6 columns

Details

m2.price - price per square meter

surface - apartment area in square meters

n.rooms - number of rooms (correlated with surface)

district - district in which apartment is located, factor with 10 levels

floor - floor

construction.date - construction year