Exract info from model

This generic function let user extract base information about model. The function returns a named list of class model_info that contain about package of model, version and task type. For wrappers like mlr or caret both, package and wrapper inforamtion are stored

model_info(model, ...)

# S3 method for lm
model_info(model, ...)

# S3 method for randomForest
model_info(model, ...)

# S3 method for svm
model_info(model, ...)

# S3 method for glm
model_info(model, ...)

# S3 method for glmnet
model_info(model, ...)

# S3 method for cv.glmnet
model_info(model, ...)

# S3 method for ranger
model_info(model, ...)

# S3 method for gbm
model_info(model, ...)

# S3 method for model_fit
model_info(model, ...)

# S3 method for train
model_info(model, ...)

# S3 method for default
model_info(model, ...)

Arguments

model
- model object

...
- another arguments

Currently supported packages are:

class `cv.glmnet` and `glmnet` - models created with `glmnet` package

class `glm` - generalized linear models

class `model_fit` - models created with `parsnip` package

class `lm` - linear models created with `stats::lm`

class `ranger` - models created with `ranger` package

class `randomForest` - random forest models created with `randomForest` package

class `svm` - support vector machines models created with the `e1071` package

class `train` - models created with `caret` package

class `gbm` - models created with `gbm` package

Value

A named list of class model_info

Examples

aps_lm_model4 <- lm(m2.price ~., data = apartments)
model_info(aps_lm_model4)

#> Package: stats 
#> Package version: 3.6.1 
#> Task type: regression

model_regr_rf <- randomForest::randomForest(m2.price~., data = apartments, ntree = 50)
model_info(model_regr_rf)

#> Package: randomForest 
#> Package version: 4.6.14 
#> Task type: regression