Plot Variable Profile Explanations

Plot Variable Profile Explanations

# S3 method for predict_profile
plot(x, ...)

Arguments

x

an object of the class predict_profile

...

other parameters

Value

An object of the class ggplot.

Plot options

ceteris_paribus

color a character. Either name of a color or name of a variable that should be used for coloring

size a numeric. Size of lines to be plotted

alpha a numeric between 0 and 1. Opacity of lines

facet_ncol number of columns for the facet_wrap

variables if not NULL then only variables will be presented

variable_type a character. If numerical then only numerical variables will be plotted. If categorical then only categorical variables will be plotted.

title a character. Plot title. By default "Ceteris Paribus profile".

subtitle a character. Plot subtitle. By default NULL - then subtitle is set to "created for the XXX, YYY model", where XXX, YYY are labels of given explainers.

categorical_type a character. How categorical variables shall be plotted? Either "lines" (default) or "bars".