Loss from Variable Dropout

Loss from Variable Dropout

variable_dropout(explainer, loss_function = function(observed, predicted)
  sum((observed - predicted)^2), ..., type = "raw", n_sample = 1000)

Arguments

explainer
a model to be explained, preprocessed by the 'explain' function

loss_function
a function thet will be used to assess variable importance

...
other parameters

type
character, type of transformation that should be applied for dropout loss. 'raw' results raw drop lossess, 'ratio' returns drop_loss/drop_loss_full_model while 'difference' returns drop_loss - drop_loss_full_model

n_sample
number of observations that should be sampled for calculations of variable leverages

Value

An object of the class 'variable_leverage_explainer'. It's a data frame with calculated average response.

Examples

library("randomForest")
HR_rf_model <- randomForest(left~., data = breakDown::HR_data, ntree = 100)

#> Warning: The response has five or fewer unique values.  Are you sure you want to do regression?

explainer_rf  <- explain(HR_rf_model, data = HR_data, y = HR_data$left)
vd_rf <- variable_dropout(explainer_rf, type = "raw")
vd_rf

#>                 variable dropout_loss        label
#> 1           _full_model_     2.864791 randomForest
#> 2  promotion_last_5years     2.864711 randomForest
#> 3                   left     2.864791 randomForest
#> 4          Work_accident     4.827576 randomForest
#> 5                 salary     6.953784 randomForest
#> 6                  sales     9.333851 randomForest
#> 7     time_spend_company    71.930029 randomForest
#> 8   average_montly_hours    89.684090 randomForest
#> 9        last_evaluation    91.273031 randomForest
#> 10        number_project   103.508723 randomForest
#> 11    satisfaction_level   155.238445 randomForest
#> 12            _baseline_   336.632141 randomForest

HR_glm_model <- glm(left~., data = breakDown::HR_data, family = "binomial")
explainer_glm <- explain(HR_glm_model, data = HR_data, y = HR_data$left)
logit <- function(x) exp(x)/(1+exp(x))
vd_glm <- variable_dropout(explainer_glm, type = "raw",
                        loss_function = function(observed, predicted) sum((observed - logit(predicted))^2))
vd_glm

#>                 variable dropout_loss label
#> 1           _full_model_     138.6548    lm
#> 2   average_montly_hours     137.9145    lm
#> 3        last_evaluation     138.5597    lm
#> 4                   left     138.6548    lm
#> 5                  sales     140.1722    lm
#> 6  promotion_last_5years     141.5771    lm
#> 7     time_spend_company     142.9918    lm
#> 8          Work_accident     146.2993    lm
#> 9                 salary     149.6998    lm
#> 10        number_project     150.8737    lm
#> 11    satisfaction_level     206.2383    lm
#> 12            _baseline_     231.1264    lm

library("xgboost")
model_martix_train <- model.matrix(left~.-1, breakDown::HR_data)
data_train <- xgb.DMatrix(model_martix_train, label = breakDown::HR_data$left)
param <- list(max_depth = 2, eta = 1, silent = 1, nthread = 2,
              objective = "binary:logistic", eval_metric = "auc")
HR_xgb_model <- xgb.train(param, data_train, nrounds = 50)
explainer_xgb <- explain(HR_xgb_model, data = model_martix_train, y = HR_data$left, label = "xgboost")
vd_xgb <- variable_dropout(explainer_xgb, type = "raw")
vd_xgb

#>                 variable dropout_loss   label
#> 1           _full_model_     20.23229 xgboost
#> 2  promotion_last_5years     20.16094 xgboost
#> 3             salessales     20.17215 xgboost
#> 4        salesaccounting     20.23229 xgboost
#> 5                salesIT     20.23229 xgboost
#> 6        salesmanagement     20.23229 xgboost
#> 7         salesmarketing     20.23229 xgboost
#> 8           salessupport     20.28994 xgboost
#> 9             salesRandD     20.30007 xgboost
#> 10      salesproduct_mng     20.43981 xgboost
#> 11         Work_accident     20.51868 xgboost
#> 12               saleshr     20.58113 xgboost
#> 13          salarymedium     20.71848 xgboost
#> 14        salestechnical     20.96317 xgboost
#> 15             salarylow     23.34276 xgboost
#> 16  average_montly_hours     42.69147 xgboost
#> 17       last_evaluation     49.73818 xgboost
#> 18        number_project     51.77436 xgboost
#> 19    time_spend_company     56.53983 xgboost
#> 20    satisfaction_level    162.02565 xgboost
#> 21            _baseline_    325.04431 xgboost