Calculate Variable Importance Explanations as Loss from Variable Dropout

From DALEX version 0.5 this function calls the feature_importance Find information how to use this function here: https://pbiecek.github.io/PM_VEE/featureImportance.html.

variable_importance(
  explainer,
  loss_function = loss_sum_of_squares,
  ...,
  type = "raw",
  n_sample = 1000
)

Arguments

explainer
a model to be explained, preprocessed by the 'explain' function

loss_function
a function thet will be used to assess variable importance

...
other parameters

type
character, type of transformation that should be applied for dropout loss. 'raw' results raw drop lossess, 'ratio' returns drop_loss/drop_loss_full_model while 'difference' returns drop_loss - drop_loss_full_model

n_sample
number of observations that should be sampled for calculation of variable importance. If negative then variable importance will be calculated on whole dataset (no sampling).

Value

An object of the class 'feature_importance'. It's a data frame with calculated average response.

References

Predictive Models: Visual Exploration, Explanation and Debugging https://pbiecek.github.io/PM_VEE/

Examples

 # \dontrun{
library(DALEX)
library("randomForest")
HR_rf_model <- randomForest(as.factor(status == "fired")~., data = HR, ntree = 100)
explainer_rf  <- explain(HR_rf_model, data = HR, y = HR$status == "fired")

#> Preparation of a new explainer is initiated
#>   -> model label       :  randomForest  (  default  )
#>   -> data              :  7847  rows  6  cols 
#>   -> target variable   :  7847  values 
#>   -> predict function  :  yhat.randomForest  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0 , mean =  0.3569096 , max =  1  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.74 , mean =  0.006923665 , max =  0.69  
#>   -> model_info        :  package randomForest , ver. 4.6.14 , task classification (  default  ) 
#>   A new explainer has been created!  

vd_rf <- variable_importance(explainer_rf, type = "raw")
vd_rf

#>        variable permutation dropout_loss        label
#> 1  _full_model_           0     41.76964 randomForest
#> 2        status           0     41.76964 randomForest
#> 3    evaluation           0    100.93660 randomForest
#> 4        gender           0    109.87826 randomForest
#> 5           age           0    141.22331 randomForest
#> 6        salary           0    159.75282 randomForest
#> 7         hours           0    234.52464 randomForest
#> 8    _baseline_           0    382.57564 randomForest
#> 9  _full_model_           1     37.17240 randomForest
#> 10       gender           1     88.96940 randomForest
#> 11          age           1    127.53740 randomForest
#> 12        hours           1    226.77990 randomForest
#> 13   evaluation           1     87.89400 randomForest
#> 14       salary           1    157.71420 randomForest
#> 15       status           1     37.17240 randomForest
#> 16   _baseline_           1    393.73240 randomForest
#> 17 _full_model_           2     43.23610 randomForest
#> 18       gender           2    115.24630 randomForest
#> 19          age           2    138.78750 randomForest
#> 20        hours           2    242.63740 randomForest
#> 21   evaluation           2    102.04350 randomForest
#> 22       salary           2    158.08390 randomForest
#> 23       status           2     43.23610 randomForest
#> 24   _baseline_           2    390.47610 randomForest
#> 25 _full_model_           3     43.01690 randomForest
#> 26       gender           3    120.08420 randomForest
#> 27          age           3    144.70360 randomForest
#> 28        hours           3    230.72280 randomForest
#> 29   evaluation           3    110.56340 randomForest
#> 30       salary           3    157.78910 randomForest
#> 31       status           3     43.01690 randomForest
#> 32   _baseline_           3    372.53690 randomForest
#> 33 _full_model_           4     45.46120 randomForest
#> 34       gender           4    122.78680 randomForest
#> 35          age           4    151.17180 randomForest
#> 36        hours           4    236.61900 randomForest
#> 37   evaluation           4    102.00630 randomForest
#> 38       salary           4    181.19600 randomForest
#> 39       status           4     45.46120 randomForest
#> 40   _baseline_           4    365.04120 randomForest
#> 41 _full_model_           5     36.28410 randomForest
#> 42       gender           5    104.16530 randomForest
#> 43          age           5    124.26600 randomForest
#> 44        hours           5    220.42540 randomForest
#> 45   evaluation           5     88.91780 randomForest
#> 46       salary           5    145.38730 randomForest
#> 47       status           5     36.28410 randomForest
#> 48   _baseline_           5    355.46410 randomForest
#> 49 _full_model_           6     42.42490 randomForest
#> 50       gender           6    105.58730 randomForest
#> 51          age           6    148.02340 randomForest
#> 52        hours           6    231.19580 randomForest
#> 53   evaluation           6    102.22850 randomForest
#> 54       salary           6    153.16380 randomForest
#> 55       status           6     42.42490 randomForest
#> 56   _baseline_           6    397.38490 randomForest
#> 57 _full_model_           7     39.32180 randomForest
#> 58       gender           7    111.24530 randomForest
#> 59          age           7    132.41920 randomForest
#> 60        hours           7    231.60440 randomForest
#> 61   evaluation           7     97.12540 randomForest
#> 62       salary           7    169.21830 randomForest
#> 63       status           7     39.32180 randomForest
#> 64   _baseline_           7    386.22180 randomForest
#> 65 _full_model_           8     40.80940 randomForest
#> 66       gender           8    100.08670 randomForest
#> 67          age           8    143.74580 randomForest
#> 68        hours           8    229.86080 randomForest
#> 69   evaluation           8     97.56510 randomForest
#> 70       salary           8    158.77750 randomForest
#> 71       status           8     40.80940 randomForest
#> 72   _baseline_           8    391.18940 randomForest
#> 73 _full_model_           9     47.11480 randomForest
#> 74       gender           9    119.53050 randomForest
#> 75          age           9    156.49260 randomForest
#> 76        hours           9    251.14820 randomForest
#> 77   evaluation           9    116.21320 randomForest
#> 78       salary           9    158.04130 randomForest
#> 79       status           9     47.11480 randomForest
#> 80   _baseline_           9    381.81480 randomForest
#> 81 _full_model_          10     42.85480 randomForest
#> 82       gender          10    111.08080 randomForest
#> 83          age          10    145.08580 randomForest
#> 84        hours          10    244.25270 randomForest
#> 85   evaluation          10    104.80880 randomForest
#> 86       salary          10    158.15680 randomForest
#> 87       status          10     42.85480 randomForest
#> 88   _baseline_          10    391.89480 randomForest

HR_glm_model <- glm(as.factor(status == "fired")~., data = HR, family = "binomial")
explainer_glm <- explain(HR_glm_model, data = HR, y = HR$status == "fired")

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  7847  rows  6  cols 
#>   -> target variable   :  7847  values 
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.00861694 , mean =  0.3638333 , max =  0.7822214  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.7755901 , mean =  -1.293796e-13 , max =  0.9820537  
#>   -> model_info        :  package stats , ver. 3.6.1 , task regression (  default  ) 
#>   A new explainer has been created!  

logit <- function(x) exp(x)/(1+exp(x))
vd_glm <- variable_importance(explainer_glm, type = "raw",
                        loss_function = function(observed, predicted)
                                     sum((observed - logit(predicted))^2))
vd_glm

#>        variable permutation dropout_loss label
#> 1  _full_model_           0     258.4282    lm
#> 2        salary           0     258.4238    lm
#> 3           age           0     258.4257    lm
#> 4        status           0     258.4282    lm
#> 5        gender           0     258.4470    lm
#> 6    evaluation           0     261.3087    lm
#> 7         hours           0     281.4606    lm
#> 8    _baseline_           0     286.7014    lm
#> 9  _full_model_           1     259.1452    lm
#> 10       gender           1     259.2419    lm
#> 11          age           1     259.1541    lm
#> 12        hours           1     284.0400    lm
#> 13   evaluation           1     260.5347    lm
#> 14       salary           1     259.1199    lm
#> 15       status           1     259.1452    lm
#> 16   _baseline_           1     287.8232    lm
#> 17 _full_model_           2     253.7079    lm
#> 18       gender           2     253.7485    lm
#> 19          age           2     253.6923    lm
#> 20        hours           2     278.7608    lm
#> 21   evaluation           2     256.1560    lm
#> 22       salary           2     253.6976    lm
#> 23       status           2     253.7079    lm
#> 24   _baseline_           2     282.3140    lm
#> 25 _full_model_           3     257.4809    lm
#> 26       gender           3     257.5091    lm
#> 27          age           3     257.4876    lm
#> 28        hours           3     279.0031    lm
#> 29   evaluation           3     261.5541    lm
#> 30       salary           3     257.4538    lm
#> 31       status           3     257.4809    lm
#> 32   _baseline_           3     289.3067    lm
#> 33 _full_model_           4     260.6942    lm
#> 34       gender           4     260.8145    lm
#> 35          age           4     260.6510    lm
#> 36        hours           4     280.9057    lm
#> 37   evaluation           4     263.5866    lm
#> 38       salary           4     260.7561    lm
#> 39       status           4     260.6942    lm
#> 40   _baseline_           4     287.1395    lm
#> 41 _full_model_           5     261.3595    lm
#> 42       gender           5     261.3248    lm
#> 43          age           5     261.3830    lm
#> 44        hours           5     284.8542    lm
#> 45   evaluation           5     266.2467    lm
#> 46       salary           5     261.3048    lm
#> 47       status           5     261.3595    lm
#> 48   _baseline_           5     285.8004    lm
#> 49 _full_model_           6     257.0093    lm
#> 50       gender           6     257.0038    lm
#> 51          age           6     257.0109    lm
#> 52        hours           6     281.7537    lm
#> 53   evaluation           6     259.9046    lm
#> 54       salary           6     257.0364    lm
#> 55       status           6     257.0093    lm
#> 56   _baseline_           6     285.3527    lm
#> 57 _full_model_           7     257.9950    lm
#> 58       gender           7     257.9382    lm
#> 59          age           7     257.9567    lm
#> 60        hours           7     279.8604    lm
#> 61   evaluation           7     260.7751    lm
#> 62       salary           7     257.9063    lm
#> 63       status           7     257.9950    lm
#> 64   _baseline_           7     284.5591    lm
#> 65 _full_model_           8     258.8784    lm
#> 66       gender           8     258.9457    lm
#> 67          age           8     258.8955    lm
#> 68        hours           8     281.1779    lm
#> 69   evaluation           8     260.6615    lm
#> 70       salary           8     258.8827    lm
#> 71       status           8     258.8784    lm
#> 72   _baseline_           8     289.5499    lm
#> 73 _full_model_           9     258.9713    lm
#> 74       gender           9     258.9337    lm
#> 75          age           9     258.9509    lm
#> 76        hours           9     284.2877    lm
#> 77   evaluation           9     260.1911    lm
#> 78       salary           9     259.0331    lm
#> 79       status           9     258.9713    lm
#> 80   _baseline_           9     286.1805    lm
#> 81 _full_model_          10     259.0405    lm
#> 82       gender          10     259.0094    lm
#> 83          age          10     259.0753    lm
#> 84        hours          10     279.9629    lm
#> 85   evaluation          10     263.4765    lm
#> 86       salary          10     259.0472    lm
#> 87       status          10     259.0405    lm
#> 88   _baseline_          10     288.9881    lm

library("xgboost")
model_martix_train <- model.matrix(status == "fired" ~ .-1, HR)
data_train <- xgb.DMatrix(model_martix_train, label = HR$status == "fired")
param <- list(max_depth = 2, eta = 1, silent = 1, nthread = 2,
              objective = "binary:logistic", eval_metric = "auc")
HR_xgb_model <- xgb.train(param, data_train, nrounds = 50)
explainer_xgb <- explain(HR_xgb_model, data = model_martix_train,
                     y = HR$status == "fired", label = "xgboost")

#> Preparation of a new explainer is initiated
#>   -> model label       :  xgboost 
#>   -> data              :  7847  rows  6  cols 
#>   -> target variable   :  7847  values 
#>   -> predict function  : yhat.default will be used (  default  )
#>   -> predicted values  :  numerical, min =  1.687903e-06 , mean =  0.363713 , max =  0.9996712  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.9885727 , mean =  0.0001203494 , max =  0.9970635  
#>   -> model_info        :  package Model of class: xgb.Booster package unrecognized , ver. Unknown , task regression (  default  ) 
#>   A new explainer has been created!  

vd_xgb <- variable_importance(explainer_xgb, type = "raw")
vd_xgb

#>        variable permutation dropout_loss   label
#> 1  _full_model_           0    100.97363 xgboost
#> 2    gendermale           0    100.97363 xgboost
#> 3    evaluation           0    115.35822 xgboost
#> 4  genderfemale           0    158.40170 xgboost
#> 5           age           0    164.95427 xgboost
#> 6        salary           0    173.90679 xgboost
#> 7         hours           0    244.47729 xgboost
#> 8    _baseline_           0    365.52545 xgboost
#> 9  _full_model_           1    108.26494 xgboost
#> 10 genderfemale           1    166.51667 xgboost
#> 11   gendermale           1    108.26494 xgboost
#> 12          age           1    166.88516 xgboost
#> 13        hours           1    240.09920 xgboost
#> 14   evaluation           1    121.87649 xgboost
#> 15       salary           1    179.65218 xgboost
#> 16   _baseline_           1    365.99301 xgboost
#> 17 _full_model_           2    100.46233 xgboost
#> 18 genderfemale           2    154.57483 xgboost
#> 19   gendermale           2    100.46233 xgboost
#> 20          age           2    161.22867 xgboost
#> 21        hours           2    253.09020 xgboost
#> 22   evaluation           2    116.68716 xgboost
#> 23       salary           2    177.31644 xgboost
#> 24   _baseline_           2    374.87524 xgboost
#> 25 _full_model_           3    111.93708 xgboost
#> 26 genderfemale           3    180.84931 xgboost
#> 27   gendermale           3    111.93708 xgboost
#> 28          age           3    183.81939 xgboost
#> 29        hours           3    256.24900 xgboost
#> 30   evaluation           3    125.14850 xgboost
#> 31       salary           3    185.83464 xgboost
#> 32   _baseline_           3    371.27604 xgboost
#> 33 _full_model_           4     94.90846 xgboost
#> 34 genderfemale           4    152.61192 xgboost
#> 35   gendermale           4     94.90846 xgboost
#> 36          age           4    152.44962 xgboost
#> 37        hours           4    232.02912 xgboost
#> 38   evaluation           4    113.99401 xgboost
#> 39       salary           4    182.55946 xgboost
#> 40   _baseline_           4    367.28781 xgboost
#> 41 _full_model_           5     95.98169 xgboost
#> 42 genderfemale           5    150.03256 xgboost
#> 43   gendermale           5     95.98169 xgboost
#> 44          age           5    167.57465 xgboost
#> 45        hours           5    242.74541 xgboost
#> 46   evaluation           5    104.76312 xgboost
#> 47       salary           5    156.43255 xgboost
#> 48   _baseline_           5    363.45684 xgboost
#> 49 _full_model_           6     97.62927 xgboost
#> 50 genderfemale           6    156.00427 xgboost
#> 51   gendermale           6     97.62927 xgboost
#> 52          age           6    164.69964 xgboost
#> 53        hours           6    245.36334 xgboost
#> 54   evaluation           6    116.02958 xgboost
#> 55       salary           6    172.83240 xgboost
#> 56   _baseline_           6    365.82481 xgboost
#> 57 _full_model_           7     97.38328 xgboost
#> 58 genderfemale           7    158.39516 xgboost
#> 59   gendermale           7     97.38328 xgboost
#> 60          age           7    154.56435 xgboost
#> 61        hours           7    248.24239 xgboost
#> 62   evaluation           7    109.38396 xgboost
#> 63       salary           7    169.87290 xgboost
#> 64   _baseline_           7    371.05115 xgboost
#> 65 _full_model_           8    101.03019 xgboost
#> 66 genderfemale           8    161.25667 xgboost
#> 67   gendermale           8    101.03019 xgboost
#> 68          age           8    169.89728 xgboost
#> 69        hours           8    238.72604 xgboost
#> 70   evaluation           8    118.62153 xgboost
#> 71       salary           8    175.65289 xgboost
#> 72   _baseline_           8    330.58521 xgboost
#> 73 _full_model_           9    103.02672 xgboost
#> 74 genderfemale           9    155.56565 xgboost
#> 75   gendermale           9    103.02672 xgboost
#> 76          age           9    165.44346 xgboost
#> 77        hours           9    253.63567 xgboost
#> 78   evaluation           9    117.55907 xgboost
#> 79       salary           9    170.99225 xgboost
#> 80   _baseline_           9    369.66208 xgboost
#> 81 _full_model_          10     99.11237 xgboost
#> 82 genderfemale          10    148.20996 xgboost
#> 83   gendermale          10     99.11237 xgboost
#> 84          age          10    162.98049 xgboost
#> 85        hours          10    234.59254 xgboost
#> 86   evaluation          10    109.51875 xgboost
#> 87       salary          10    167.92220 xgboost
#> 88   _baseline_          10    375.24234 xgboost

 # }