Marginal Response for a Single Variable

Calculates the average model response as a function of a single selected variable. Use the 'type' parameter to select the type of marginal response to be calculated. Currently for numeric variables we have Partial Dependency and Accumulated Local Effects implemented. Current implementation uses the 'pdp' package (Brandon M. Greenwell (2017). pdp: An R Package for Constructing Partial Dependence Plots. The R Journal, 9(1), 421--436.) and 'ALEPlot' (Dan Apley (2017). ALEPlot: Accumulated Local Effects Plots and Partial Dependence Plots.)

variable_response(explainer, variable, type = "pdp", trans = explainer$link,
  ...)

Arguments

explainer
a model to be explained, preprocessed by the 'explain' function

variable
character - name of a single variable

type
character - type of the response to be calculated. Currently following options are implemented: 'pdp' for Partial Dependency and 'ale' for Accumulated Local Effects

trans
function - a transformation/link function that shall be applied to raw model predictions. This will be inherited from the explainer.

...
other parameters

Value

An object of the class 'svariable_response_explainer'. It's a data frame with calculated average response.

Details

For factor variables we are using the 'factorMerger' package. Please note that the argument type must be set to 'factor' to use this method.

Examples

library("breakDown")
logit <- function(x) exp(x)/(1+exp(x))

HR_glm_model <- glm(left~., data = breakDown::HR_data, family = "binomial")
explainer_glm <- explain(HR_glm_model, data = HR_data, trans=logit)
expl_glm <- variable_response(explainer_glm, "satisfaction_level", "pdp")
expl_glm

#>         x            y                var type label
#> 1  0.0900 0.560819.... satisfaction_level  pdp    lm
#> 2  0.1082 0.485549.... satisfaction_level  pdp    lm
#> 3  0.1264 0.410280.... satisfaction_level  pdp    lm
#> 4  0.1446 0.335010.... satisfaction_level  pdp    lm
#> 5  0.1628 0.259741.... satisfaction_level  pdp    lm
#> 6  0.1810 0.184471.... satisfaction_level  pdp    lm
#> 7  0.1992 0.109202.... satisfaction_level  pdp    lm
#> 8  0.2174 0.033932.... satisfaction_level  pdp    lm
#> 9  0.2356 -0.04133.... satisfaction_level  pdp    lm
#> 10 0.2538 -0.11660.... satisfaction_level  pdp    lm
#> 11 0.2720 -0.19187.... satisfaction_level  pdp    lm
#> 12 0.2902 -0.26714.... satisfaction_level  pdp    lm
#> 13 0.3084 -0.34241.... satisfaction_level  pdp    lm
#> 14 0.3266 -0.41768.... satisfaction_level  pdp    lm
#> 15 0.3448 -0.49295.... satisfaction_level  pdp    lm
#> 16 0.3630 -0.56822.... satisfaction_level  pdp    lm
#> 17 0.3812 -0.64349.... satisfaction_level  pdp    lm
#> 18 0.3994 -0.71876.... satisfaction_level  pdp    lm
#> 19 0.4176 -0.79403.... satisfaction_level  pdp    lm
#> 20 0.4358 -0.86930.... satisfaction_level  pdp    lm
#> 21 0.4540 -0.94457.... satisfaction_level  pdp    lm
#> 22 0.4722 -1.01984.... satisfaction_level  pdp    lm
#> 23 0.4904 -1.09511.... satisfaction_level  pdp    lm
#> 24 0.5086 -1.17038.... satisfaction_level  pdp    lm
#> 25 0.5268 -1.24564.... satisfaction_level  pdp    lm
#> 26 0.5450 -1.32091.... satisfaction_level  pdp    lm
#> 27 0.5632 -1.39618.... satisfaction_level  pdp    lm
#> 28 0.5814 -1.47145.... satisfaction_level  pdp    lm
#> 29 0.5996 -1.54672.... satisfaction_level  pdp    lm
#> 30 0.6178 -1.62199.... satisfaction_level  pdp    lm
#> 31 0.6360 -1.69726.... satisfaction_level  pdp    lm
#> 32 0.6542 -1.77253.... satisfaction_level  pdp    lm
#> 33 0.6724 -1.84780.... satisfaction_level  pdp    lm
#> 34 0.6906 -1.92307.... satisfaction_level  pdp    lm
#> 35 0.7088 -1.99834.... satisfaction_level  pdp    lm
#> 36 0.7270 -2.07361.... satisfaction_level  pdp    lm
#> 37 0.7452 -2.14888.... satisfaction_level  pdp    lm
#> 38 0.7634 -2.22415.... satisfaction_level  pdp    lm
#> 39 0.7816 -2.29942.... satisfaction_level  pdp    lm
#> 40 0.7998 -2.37469.... satisfaction_level  pdp    lm
#> 41 0.8180 -2.44996.... satisfaction_level  pdp    lm
#> 42 0.8362 -2.52523.... satisfaction_level  pdp    lm
#> 43 0.8544 -2.60050.... satisfaction_level  pdp    lm
#> 44 0.8726 -2.67577.... satisfaction_level  pdp    lm
#> 45 0.8908 -2.75104.... satisfaction_level  pdp    lm
#> 46 0.9090 -2.82630.... satisfaction_level  pdp    lm
#> 47 0.9272 -2.90157.... satisfaction_level  pdp    lm
#> 48 0.9454 -2.97684.... satisfaction_level  pdp    lm
#> 49 0.9636 -3.05211.... satisfaction_level  pdp    lm
#> 50 0.9818 -3.12738.... satisfaction_level  pdp    lm
#> 51 1.0000 -3.20265.... satisfaction_level  pdp    lm

#\dontrun{
library("randomForest")
HR_rf_model <- randomForest(factor(left)~., data = breakDown::HR_data, ntree = 100)
explainer_rf  <- explain(HR_rf_model, data = HR_data,
                       predict_function = function(model, x)
                              predict(model, x, type = "prob")[,2])
expl_rf  <- variable_response(explainer_rf, variable = "satisfaction_level", type = "pdp",
                       which.class = 2, prob = TRUE)
expl_rf

#>         x            y                var type        label
#> 1  0.0900 0.714120.... satisfaction_level  pdp randomForest
#> 2  0.1082 0.682272.... satisfaction_level  pdp randomForest
#> 3  0.1264 0.142329.... satisfaction_level  pdp randomForest
#> 4  0.1446 0.138485.... satisfaction_level  pdp randomForest
#> 5  0.1628 0.096517.... satisfaction_level  pdp randomForest
#> 6  0.1810 0.086972.... satisfaction_level  pdp randomForest
#> 7  0.1992 0.084976.... satisfaction_level  pdp randomForest
#> 8  0.2174 0.080126.... satisfaction_level  pdp randomForest
#> 9  0.2356 0.081410.... satisfaction_level  pdp randomForest
#> 10 0.2538 0.086608.... satisfaction_level  pdp randomForest
#> 11 0.2720 0.089170.... satisfaction_level  pdp randomForest
#> 12 0.2902 0.094750.... satisfaction_level  pdp randomForest
#> 13 0.3084 0.146443.... satisfaction_level  pdp randomForest
#> 14 0.3266 0.134232.... satisfaction_level  pdp randomForest
#> 15 0.3448 0.141832.... satisfaction_level  pdp randomForest
#> 16 0.3630 0.188340.... satisfaction_level  pdp randomForest
#> 17 0.3812 0.184614.... satisfaction_level  pdp randomForest
#> 18 0.3994 0.183676.... satisfaction_level  pdp randomForest
#> 19 0.4176 0.194706.... satisfaction_level  pdp randomForest
#> 20 0.4358 0.184025.... satisfaction_level  pdp randomForest
#> 21 0.4540 0.185971.... satisfaction_level  pdp randomForest
#> 22 0.4722 0.096976.... satisfaction_level  pdp randomForest
#> 23 0.4904 0.060362.... satisfaction_level  pdp randomForest
#> 24 0.5086 0.056032.... satisfaction_level  pdp randomForest
#> 25 0.5268 0.053508.... satisfaction_level  pdp randomForest
#> 26 0.5450 0.054545.... satisfaction_level  pdp randomForest
#> 27 0.5632 0.052826.... satisfaction_level  pdp randomForest
#> 28 0.5814 0.051665.... satisfaction_level  pdp randomForest
#> 29 0.5996 0.051348.... satisfaction_level  pdp randomForest
#> 30 0.6178 0.051652.... satisfaction_level  pdp randomForest
#> 31 0.6360 0.050426.... satisfaction_level  pdp randomForest
#> 32 0.6542 0.051118.... satisfaction_level  pdp randomForest
#> 33 0.6724 0.054053.... satisfaction_level  pdp randomForest
#> 34 0.6906 0.056022.... satisfaction_level  pdp randomForest
#> 35 0.7088 0.071012.... satisfaction_level  pdp randomForest
#> 36 0.7270 0.100087.... satisfaction_level  pdp randomForest
#> 37 0.7452 0.099829.... satisfaction_level  pdp randomForest
#> 38 0.7634 0.099741.... satisfaction_level  pdp randomForest
#> 39 0.7816 0.103230.... satisfaction_level  pdp randomForest
#> 40 0.7998 0.103267.... satisfaction_level  pdp randomForest
#> 41 0.8180 0.101089.... satisfaction_level  pdp randomForest
#> 42 0.8362 0.100384.... satisfaction_level  pdp randomForest
#> 43 0.8544 0.102093.... satisfaction_level  pdp randomForest
#> 44 0.8726 0.099193.... satisfaction_level  pdp randomForest
#> 45 0.8908 0.103068.... satisfaction_level  pdp randomForest
#> 46 0.9090 0.096130.... satisfaction_level  pdp randomForest
#> 47 0.9272 0.059521.... satisfaction_level  pdp randomForest
#> 48 0.9454 0.052171.... satisfaction_level  pdp randomForest
#> 49 0.9636 0.051826.... satisfaction_level  pdp randomForest
#> 50 0.9818 0.051826.... satisfaction_level  pdp randomForest
#> 51 1.0000 0.051826.... satisfaction_level  pdp randomForest

#}