Wrap Various Predict Functions

This function is a wrapper over various predict functions for different models and differnt model structures. The wrapper returns a single numeric score for each new observation. To do this it uses different extraction techniques for models from different classes, like for classification random forest is forces the output to be probabilities not classes itself.

yhat(X.model, newdata, ...)

# S3 method for lm
yhat(X.model, newdata, ...)

# S3 method for randomForest
yhat(X.model, newdata, ...)

# S3 method for svm
yhat(X.model, newdata, ...)

# S3 method for gbm
yhat(X.model, newdata, ...)

# S3 method for glm
yhat(X.model, newdata, ...)

# S3 method for cv.glmnet
yhat(X.model, newdata, ...)

# S3 method for glmnet
yhat(X.model, newdata, ...)

# S3 method for ranger
yhat(X.model, newdata, ...)

# S3 method for model_fit
yhat(X.model, newdata, ...)

# S3 method for train
yhat(X.model, newdata, ...)

# S3 method for lrm
yhat(X.model, newdata, ...)

# S3 method for rpart
yhat(X.model, newdata, ...)

# S3 method for default
yhat(X.model, newdata, ...)

Arguments

X.model
object - a model to be explained

newdata
data.frame or matrix - observations for prediction

...
other parameters that will be passed to the predict function

Value

An numeric matrix of predictions

Details

Currently supported packages are:

class cv.glmnet and glmnet - models created with glmnet package,

class glm - generalized linear models created with glm,

class model_fit - models created with parsnip package,

class lm - linear models created with lm,

class ranger - models created with ranger package,

class randomForest - random forest models created with randomForest package,

class svm - support vector machines models created with the e1071 package,

class train - models created with caret package,

class gbm - models created with gbm package,

class lrm - models created with rms package,

class rpart - models created with rpart package.