Human Resources Data

Datasets HR and HR_test are artificial, generated form the same model. Structure of the dataset is based on a real data, from Human Resources department with information which employees were promoted, which were fired.

data(HR)

Format

a data frame with 10000 rows and 6 columns

Details

Values are generated in a way to: - have interaction between age and gender for the 'fired' variable - have non monotonic relation for the salary variable - have linear effects for hours and evaluation.

gender - gender of an employee.

age - age of an employee in the moment of evaluation.

hours - average number of working hours per week.

evaluation - evaluation in the scale 2 (bad) - 5 (very good).

salary - level of salary in the scale 0 (lowest) - 5 (highest).

status - target variable, either `fired` or `promoted` or `ok`.