Predictions for the Explainer

This is a generic predict() function works for explainer objects.

# S3 method for explainer
predict(object, newdata, ...)

model_prediction(explainer, new_data, ...)

Arguments

object
a model to be explained, object of the class explainer

newdata
data.frame or matrix - observations for prediction

...
other parameters that will be passed to the predict function

explainer
a model to be explained, object of the class explainer

new_data
data.frame or matrix - observations for prediction

Value

An numeric matrix of predictions

Examples

HR_glm_model <- glm(status == "fired"~., data = HR, family = "binomial")
explainer_glm <- explain(HR_glm_model, data = HR)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  7847  rows  6  cols 
#>   -> target variable   :  not specified! (  WARNING  )
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.00861694 , mean =  0.3638333 , max =  0.7822214  
#>   -> model_info        :  package stats , ver. 4.0.0 , task classification (  default  ) 
#>   -> residual function :  difference between y and yhat (  default  )
#>   A new explainer has been created!  

predict(explainer_glm, HR[1:3,])

#>         1         2         3 
#> 0.5139357 0.7384469 0.6412859 

 # \dontrun{
library("ranger")
HR_ranger_model <- ranger(status~., data = HR, num.trees = 50, probability = TRUE)
explainer_ranger  <- explain(HR_ranger_model, data = HR)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  7847  rows  6  cols 
#>   -> target variable   :  not specified! (  WARNING  )
#>   -> predict function  :  yhat.ranger  will be used (  default  )
#>   -> predicted values  :  predict function returns multiple columns:  3  (  WARNING  ) some of functionalities may not work 
#>   -> model_info        :  package ranger , ver. 0.12.1 , task multiclass (  default  ) 
#>   -> residual function :  difference between 1 and probability of true class (  default  )
#>   A new explainer has been created!  

predict(explainer_ranger, HR[1:3,])

#>          fired         ok    promoted
#> [1,] 0.7036873 0.28676243 0.009550242
#> [2,] 0.9742988 0.02112259 0.004578557
#> [3,] 0.9463568 0.04511011 0.008533122

model_prediction(explainer_ranger, HR[1:3,])

#>          fired         ok    promoted
#> [1,] 0.7036873 0.28676243 0.009550242
#> [2,] 0.9742988 0.02112259 0.004578557
#> [3,] 0.9463568 0.04511011 0.008533122

 # }