Feature Importance Calculated as Loss from Feature Dropout

This function is set deprecated. It is suggested to use feature_importance instead. Find information how to use these functions here: https://pbiecek.github.io/PM_VEE/featureImportance.html.

variable_importance(explainer, loss_function = loss_sum_of_squares, ...,
  type = "raw", n_sample = 1000)

Arguments

explainer
a model to be explained, preprocessed by the 'explain' function

loss_function
a function thet will be used to assess variable importance

...
other parameters

type
character, type of transformation that should be applied for dropout loss. 'raw' results raw drop lossess, 'ratio' returns drop_loss/drop_loss_full_model while 'difference' returns drop_loss - drop_loss_full_model

n_sample
number of observations that should be sampled for calculation of variable importance. If negative then variable importance will be calculated on whole dataset (no sampling).

Value

An object of the class 'variable_leverage_explainer'. It's a data frame with calculated average response.

References

Predictive Models: Visual Exploration, Explanation and Debugging https://pbiecek.github.io/PM_VEE/

Examples

 

library(DALEX)
library("randomForest")
HR_rf_model <- randomForest(as.factor(status == "fired")~., data = HR, ntree = 100)
explainer_rf  <- explain(HR_rf_model, data = HR, y = HR$status == "fired")

#> Preparation of a new explainer is initiated
#>   -> model label       :  randomForest  (default)
#>   -> data              :  7847  rows  6  cols 
#>   -> target variable   :  7847  values 
#>   -> predict function  :  yhat.randomForest  will be used (default)
#>   -> predicted values  :  numerical, min =  0 , mean =  0.3609736 , max =  1  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -0.76 , mean =  0.002859692 , max =  0.75  
#> A new explainer has been created!

vd_rf <- variable_importance(explainer_rf, type = "raw")
vd_rf

#>       variable dropout_loss        label
#> 1 _full_model_      47.7020 randomForest
#> 2       status      47.7020 randomForest
#> 3   evaluation     101.1195 randomForest
#> 4       gender     119.2739 randomForest
#> 5          age     143.8208 randomForest
#> 6       salary     163.4523 randomForest
#> 7        hours     239.3491 randomForest
#> 8   _baseline_     359.6420 randomForest

HR_glm_model <- glm(as.factor(status == "fired")~., data = HR, family = "binomial")
explainer_glm <- explain(HR_glm_model, data = HR, y = HR$status == "fired")

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (default)
#>   -> data              :  7847  rows  6  cols 
#>   -> target variable   :  7847  values 
#>   -> predict function  :  yhat.glm  will be used (default)
#>   -> predicted values  :  numerical, min =  0.00861694 , mean =  0.3638333 , max =  0.7822214  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -0.7755901 , mean =  -1.293887e-13 , max =  0.9820537  
#> A new explainer has been created!

logit <- function(x) exp(x)/(1+exp(x))
vd_glm <- variable_importance(explainer_glm, type = "raw",
                        loss_function = function(observed, predicted)
                                     sum((observed - logit(predicted))^2))
vd_glm

#>       variable dropout_loss label
#> 1 _full_model_     256.7451    lm
#> 2       gender     256.7129    lm
#> 3          age     256.7361    lm
#> 4       status     256.7451    lm
#> 5       salary     256.7822    lm
#> 6   evaluation     260.6747    lm
#> 7        hours     282.7334    lm
#> 8   _baseline_     285.0523    lm

library("xgboost")
model_martix_train <- model.matrix(status == "fired" ~ .-1, HR)
data_train <- xgb.DMatrix(model_martix_train, label = HR$status == "fired")
param <- list(max_depth = 2, eta = 1, silent = 1, nthread = 2,
              objective = "binary:logistic", eval_metric = "auc")
HR_xgb_model <- xgb.train(param, data_train, nrounds = 50)
explainer_xgb <- explain(HR_xgb_model, data = model_martix_train,
                     y = HR$status == "fired", label = "xgboost")

#> Preparation of a new explainer is initiated
#>   -> model label       :  xgboost 
#>   -> data              :  7847  rows  6  cols 
#>   -> target variable   :  7847  values 
#>   -> predict function  : yhat.default will be used (default)
#>   -> predicted values  :  numerical, min =  1.687903e-06 , mean =  0.363713 , max =  0.9996712  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -0.9885727 , mean =  0.0001203497 , max =  0.9970635  
#> A new explainer has been created!

vd_xgb <- variable_importance(explainer_xgb, type = "raw")
vd_xgb

#>       variable dropout_loss   label
#> 1 _full_model_     91.75335 xgboost
#> 2   gendermale     91.75335 xgboost
#> 3   evaluation    106.01494 xgboost
#> 4 genderfemale    149.91086 xgboost
#> 5          age    153.98590 xgboost
#> 6       salary    172.61664 xgboost
#> 7        hours    233.05302 xgboost
#> 8   _baseline_    344.54982 xgboost