Calculate Feature Importance Explanations as Loss from Feature Dropout

This function is set deprecated. It is suggested to use feature_importance instead. Find information how to use these functions here: https://pbiecek.github.io/PM_VEE/featureImportance.html.

variable_importance(explainer, loss_function = loss_sum_of_squares, ...,
  type = "raw", n_sample = 1000)

Arguments

explainer
a model to be explained, preprocessed by the 'explain' function

loss_function
a function thet will be used to assess variable importance

...
other parameters

type
character, type of transformation that should be applied for dropout loss. 'raw' results raw drop lossess, 'ratio' returns drop_loss/drop_loss_full_model while 'difference' returns drop_loss - drop_loss_full_model

n_sample
number of observations that should be sampled for calculation of variable importance. If negative then variable importance will be calculated on whole dataset (no sampling).

Value

An object of the class 'variable_leverage_explainer'. It's a data frame with calculated average response.

References

Predictive Models: Visual Exploration, Explanation and Debugging https://pbiecek.github.io/PM_VEE/

Examples

 # \dontrun{
library(DALEX)
library("randomForest")
HR_rf_model <- randomForest(as.factor(status == "fired")~., data = HR, ntree = 100)
explainer_rf  <- explain(HR_rf_model, data = HR, y = HR$status == "fired")

#> Preparation of a new explainer is initiated
#>   -> model label       :  randomForest  (  default  )
#>   -> data              :  7847  rows  6  cols 
#>   -> target variable   :  7847  values 
#>   -> predict function  :  yhat.randomForest  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0 , mean =  0.3600905 , max =  1  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.77 , mean =  0.003742832 , max =  0.71  
#>   -> model_info        :  package randomForest , ver. 4.6.14 , task classification (  default  ) 
#>   A new explainer has been created!  

vd_rf <- variable_importance(explainer_rf, type = "raw")
vd_rf

#>       variable dropout_loss        label
#> 1 _full_model_      41.0400 randomForest
#> 2       status      41.0400 randomForest
#> 3   evaluation      99.8170 randomForest
#> 4       gender     101.2611 randomForest
#> 5          age     139.6908 randomForest
#> 6       salary     148.5117 randomForest
#> 7        hours     236.7322 randomForest
#> 8   _baseline_     410.9000 randomForest

HR_glm_model <- glm(as.factor(status == "fired")~., data = HR, family = "binomial")
explainer_glm <- explain(HR_glm_model, data = HR, y = HR$status == "fired")

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  7847  rows  6  cols 
#>   -> target variable   :  7847  values 
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.00861694 , mean =  0.3638333 , max =  0.7822214  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.7755901 , mean =  -1.293796e-13 , max =  0.9820537  
#>   -> model_info        :  package stats , ver. 3.6.1 , task regression (  default  ) 
#>   A new explainer has been created!  

logit <- function(x) exp(x)/(1+exp(x))
vd_glm <- variable_importance(explainer_glm, type = "raw",
                        loss_function = function(observed, predicted)
                                     sum((observed - logit(predicted))^2))
vd_glm

#>       variable dropout_loss label
#> 1 _full_model_     257.1169    lm
#> 2       salary     257.0588    lm
#> 3       gender     257.0767    lm
#> 4       status     257.1169    lm
#> 5          age     257.1469    lm
#> 6   evaluation     259.5558    lm
#> 7        hours     281.8584    lm
#> 8   _baseline_     288.7987    lm

library("xgboost")

#> Error in library("xgboost"): there is no package called ‘xgboost’

model_martix_train <- model.matrix(status == "fired" ~ .-1, HR)
data_train <- xgb.DMatrix(model_martix_train, label = HR$status == "fired")

#> Error in xgb.DMatrix(model_martix_train, label = HR$status == "fired"): could not find function "xgb.DMatrix"

param <- list(max_depth = 2, eta = 1, silent = 1, nthread = 2,
              objective = "binary:logistic", eval_metric = "auc")
HR_xgb_model <- xgb.train(param, data_train, nrounds = 50)

#> Error in xgb.train(param, data_train, nrounds = 50): could not find function "xgb.train"

explainer_xgb <- explain(HR_xgb_model, data = model_martix_train,
                     y = HR$status == "fired", label = "xgboost")

#> Preparation of a new explainer is initiated
#>   -> model label       :  xgboost 
#>   -> data              :  7847  rows  6  cols 
#>   -> target variable   :  7847  values 

#> Error in paste0("yhat.", class(model)): object 'HR_xgb_model' not found

vd_xgb <- variable_importance(explainer_xgb, type = "raw")

#> Error in "explainer" %in% class(explainer): object 'explainer_xgb' not found

vd_xgb

#> Error in eval(expr, envir, enclos): object 'vd_xgb' not found

 # }